refactor: extract Makeup API into a standalone feature module
This commit migrates the Makeup API integration from the `cosmetics` feature into a dedicated `makeupapi` module to improve separation of concerns and reusability. It also introduces documentation and UI hooks for service discovery monitoring.

- **Module Extraction**:
    - Created the `:applications:kocolor:features:makeupapi` module.
    - Relocated `MakeupApiService`, `MakeupProductDto`, `MakeupRepository`, and Hilt DI modules (`MakeupModule`, `MakeupRepositoryModule`) from `:features:cosmetics` to the new module.
    - Updated `settings.gradle.kts` and `gradle.xml` to include the new feature module.
- **Dependency Management**:
    - Updated `mobile` app, `boxcapture`, and `cosmetics` features to depend on `:applications:kocolor:features:makeupapi`.
    - Refactored `BoxCaptureViewModel` and `CosmeticsViewModel` to use the repository from the new package location.
- **BoxCapture Enhancements**:
    - Integrated `discoveryStatus` into `BoxCaptureRoute` to track real-time API connectivity during product capture.
    - Removed unused `KoColorRoute` import.
- **Documentation**:
    - Updated `external_services.md` to include **Discovery Health Monitoring**, detailing how the app visualizes real-time connectivity for services like OBF, FDA, and PubChem.
    - Renumbered existing sections in the external services documentation for consistency.
