+++ .idea/gradle.xml
@@ -100,6 +100,7 @@
             <option value="$PROJECT_DIR$/applications/kocolor/features/cosmetics" />
             <option value="$PROJECT_DIR$/applications/kocolor/features/fda" />
             <option value="$PROJECT_DIR$/applications/kocolor/features/inventory" />
+            <option value="$PROJECT_DIR$/applications/kocolor/features/makeupapi" />
             <option value="$PROJECT_DIR$/applications/kocolor/features/products" />
             <option value="$PROJECT_DIR$/applications/kocolor/features/routines" />
             <option value="$PROJECT_DIR$/applications/kocolor/features/stitch" />
