+++ applications/kocolor/apps/mobile/build.gradle.kts
@@ -77,6 +77,7 @@ dependencies {
     implementation(project(":applications:kocolor:features:cosmetics"))
     implementation(project(":applications:kocolor:features:fda"))
     implementation(project(":applications:kocolor:features:chemicals"))
+    implementation(project(":applications:kocolor:features:makeupapi"))
     // implementation(project(":features:xr:ar:naillab"))
     // implementation(project(":features:xr:ar:facelab"))
