+++ applications/kocolor/features/cosmetics/build.gradle.kts
@@ -25,6 +25,7 @@ dependencies {
     implementation(project(":features:obf"))
     implementation(project(":applications:kocolor:features:fda"))
     implementation(project(":applications:kocolor:features:chemicals"))
+    implementation(project(":applications:kocolor:features:makeupapi"))
 
     implementation(libs.androidx.compose.material.icons.extended)
     implementation(libs.androidx.activity.compose)
