+++ applications/kocolor/features/makeupapi/src/main/java/com/zoewave/probase/kocolor/features/makeupapi/data/remote/model/MakeupProductDto.kt
@@ -1,4 +1,4 @@
-package com.zoewave.probase.kocolor.features.cosmetics.data.remote.model
+package com.zoewave.probase.kocolor.features.makeupapi.data.remote.model
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
