+++ settings.gradle.kts
@@ -166,6 +166,7 @@ include(":applications:kocolor:features:stitch")
 include(":applications:kocolor:features:store")
 include(":applications:kocolor:features:boxcapture")
 include(":applications:kocolor:features:clothingcapture")
+include(":applications:kocolor:features:makeupapi")
 // include(":applications:kocolor:features:stitch")
 include(":applications:kocolor:apps:mobile")
 include(":applications:kocolor:apps:mobile:core")
